Add CPU device plugin
CPU device plugin advertises cpu pool as a set of devices. For the
actual cpu allocation cmk is used as a backend. The advertised cpu
device ids are not real cpu numbers so there is a mapping between device ids
and (cmk) allocated cpu numbers. The mapping is stored in a file.

CMK webhook is modified to support mutating pods with different resource
requests. The default resource name can be overridden in the configmap with
'resource-name' attribute.

Signed-off-by: Timo Lindqvist <timo.lindqvist@nokia.com>